requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
tqdm>=4.65.0

# Configuration
PyYAML>=6.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# Utilities
pandas>=2.0.0
tensorboard>=2.13.0

# Optional: For faster linear algebra
# mkl>=2023.0  # Intel MKL for faster computations
